+++ packages/odata-service-inquirer/test/unit/prompts/sap-system/system-selection/questions.test.ts
@@ -3,6 +3,7 @@ import { ODataVersion } from '@sap-ux/axios-extension';
 import type { Destination, Destinations } from '@sap-ux/btp-utils';
 import { WebIDEAdditionalData, WebIDEUsage } from '@sap-ux/btp-utils';
 import type { ListQuestion } from '@sap-ux/inquirer-common';
+import { hostEnvironment } from '@sap-ux/fiori-generator-shared';
 import type { BackendSystem } from '@sap-ux/store';
 import type { ListChoiceOptions, Question } from 'inquirer';
 import { ERROR_TYPE } from '../../../../../src';
@@ -11,22 +12,22 @@ import type { ValidationResult } from '../../../../../src/prompts/connectionVali
 import { ConnectionValidator } from '../../../../../src/prompts/connectionValidator';
 import { newSystemPromptNames } from '../../../../../src/prompts/datasources/sap-system/new-system/types';
 import * as promptHelpers from '../../../../../src/prompts/datasources/sap-system/system-selection/prompt-helpers';
+import { NewSystemChoice } from '../../../../../src/prompts/datasources/sap-system/system-selection/prompt-helpers';
 import type {
     SystemSelectionAnswers,
     SystemSelectionAnswerType
 } from '../../../../../src/prompts/datasources/sap-system/system-selection/questions';
 import {
     getSystemConnectionQuestions,
-    getSystemSelectionQuestions,
-    newSystemChoiceValue
+    getSystemSelectionQuestions
 } from '../../../../../src/prompts/datasources/sap-system/system-selection/questions';
 import LoggerHelper from '../../../../../src/prompts/logger-helper';
-import { hostEnvironment, promptNames } from '../../../../../src/types';
-import { getHostEnvironment, PromptState } from '../../../../../src/utils';
+import { promptNames } from '../../../../../src/types';
+import { getPromptHostEnvironment, PromptState } from '../../../../../src/utils';
 
 jest.mock('../../../../../src/utils', () => ({
     ...jest.requireActual('../../../../../src/utils'),
-    getHostEnvironment: jest.fn()
+    getPromptHostEnvironment: jest.fn()
 }));
 
 const backendSystemBasic: BackendSystem = {
@@ -191,7 +192,7 @@ describe('Test system selection prompts', () => {
         systemChoices = (systemSelectionPrompt as ListQuestion).choices as ListChoiceOptions<SystemSelectionAnswers>[];
         expect(systemChoices[0].value as SystemSelectionAnswerType).toEqual({
             type: 'newSystemChoice',
-            system: newSystemChoiceValue
+            system: NewSystemChoice
         });
     });
 
@@ -202,7 +203,7 @@ describe('Test system selection prompts', () => {
             (question) => question.name === `systemSelection:${promptNames.serviceSelection}`
         );
         const serviceSelectionWhenResult = ((systemServicePrompt as Question).when as Function)({
-            systemSelection: { type: 'newSystemChoice', system: newSystemChoiceValue }
+            systemSelection: { type: 'newSystemChoice', system: NewSystemChoice }
         } as SystemSelectionAnswers);
         expect(serviceSelectionWhenResult).toEqual(false);
 
@@ -212,14 +213,14 @@ describe('Test system selection prompts', () => {
         );
         expect(
             ((newSystemTypePrompt as Question).when as Function)({
-                systemSelection: { type: 'newSystemChoice', system: newSystemChoiceValue }
+                systemSelection: { type: 'newSystemChoice', system: NewSystemChoice }
             })
         ).toBe(true);
     });
 
     test('getSystemConnectionQuestions: BAS (Destination)', async () => {
         const connectValidator = new ConnectionValidator();
-        (getHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
+        (getPromptHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
         mockIsAppStudio = true;
 
         const systemConnectionQuestions = await getSystemConnectionQuestions(connectValidator);
@@ -284,6 +285,14 @@ describe('Test system selection prompts', () => {
         expect(await (destServicePathPrompt.validate as Function)?.('')).toEqual(
             t('prompts.destinationServicePath.invalidServicePathWarning')
         );
+        // Invalid service path if only 1 char
+        expect(await (destServicePathPrompt.validate as Function)?.('/')).toEqual(
+            t('prompts.destinationServicePath.invalidServicePathWarning')
+        );
+        // Invalid service path if starts with double slash
+        expect(await (destServicePathPrompt.validate as Function)?.('//123')).toEqual(
+            t('prompts.destinationServicePath.invalidServicePathWarning')
+        );
         expect(
             await (destServicePathPrompt.validate as Function)?.('/\\', {
                 [promptNames.systemSelection]: {
@@ -311,7 +320,7 @@ describe('Test system selection prompts', () => {
     test('getSystemConnectionQuestions: non-BAS (BackendSystem, AuthType: basic)', async () => {
         mockIsAppStudio = false;
         const connectValidator = new ConnectionValidator();
-        (getHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
+        (getPromptHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
         const systemConnectionQuestions = await getSystemConnectionQuestions(connectValidator);
         expect(systemConnectionQuestions).toHaveLength(4);
         expect(systemConnectionQuestions[0].name).toBe('systemSelection');
@@ -385,7 +394,7 @@ describe('Test system selection prompts', () => {
     test('getSystemConnectionQuestions: non-BAS (BackendSystem, AuthType: reentranceTicket)', async () => {
         mockIsAppStudio = false;
         const connectValidator = new ConnectionValidator();
-        (getHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
+        (getPromptHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
         const connectWithBackendSystemSpy = jest.spyOn(promptHelpers, 'connectWithBackendSystem');
         backendSystems.push(backendSystemReentrance);
 
@@ -423,7 +432,7 @@ describe('Test system selection prompts', () => {
     test('getSystemConnectionQuestions: non-BAS (BackendSystem, AuthType: serviceKeys)', async () => {
         mockIsAppStudio = false;
         const connectValidator = new ConnectionValidator();
-        (getHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
+        (getPromptHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
         const connectWithBackendSystemSpy = jest.spyOn(promptHelpers, 'connectWithBackendSystem');
         backendSystems.push(backendSystemReentrance);
 
@@ -460,7 +469,7 @@ describe('Test system selection prompts', () => {
 
     test('should execute additional prompt on CLI (if autocomplete is not used) to handle YUI validate function', async () => {
         mockIsAppStudio = false;
-        (getHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
+        (getPromptHostEnvironment as jest.Mock).mockReturnValue(hostEnvironment.cli);
         const connectWithBackendSystemSpy = jest.spyOn(promptHelpers, 'connectWithBackendSystem');
         const systemConnectionQuestions = await getSystemConnectionQuestions(new ConnectionValidator());
         const validateSystemSelectionCliPrompt = systemConnectionQuestions.find(
@@ -493,5 +502,23 @@ describe('Test system selection prompts', () => {
             connectionValidatorMock,
             undefined
         );
+    });
+
+    test('Should set the default system choice based on the defaultChoice options', async () => {
+        backendSystems.push(backendSystemReentrance);
+        const defaultChoice = backendSystemReentrance.name;
+        const systemSelectionQuestions = await getSystemSelectionQuestions({
+            [promptNames.systemSelection]: { defaultChoice }
+        });
+        const systemSelectionPrompt = systemSelectionQuestions.find(
+            (question) => question.name === promptNames.systemSelection
+        );
+        const defaultIndex = (systemSelectionPrompt as Question).default;
+        expect(((systemSelectionPrompt as ListQuestion).choices as [])[defaultIndex]).toMatchObject({
+            value: {
+                system: backendSystemReentrance,
+                type: 'backendSystem'
+            }
+        });
     });
 });