+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.8.1
+
+### Patch Changes
+
+-   Updated dependencies [74dc5fe]
+    -   @sap-ux/inquirer-common@0.5.4
+
+## 0.8.0
+
+### Minor Changes
+
+-   d11283f: Remove the enableNpmWorkspaces prompt (and all references)
+
+## 0.7.16
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.3
+
+## 0.7.15
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.5.2
+
+## 0.7.14
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.28.7
+-   @sap-ux/inquirer-common@0.5.1
+
+## 0.7.13
+
+### Patch Changes
+
+-   Updated dependencies [2886db3]
+    -   @sap-ux/inquirer-common@0.5.0
+
 ## 0.7.12
 
 ### Patch Changes